abstract logger into logger service; make cli service a single class
